Fix bug - use "16-mixed" for precision instead of "16" - fix model.forward() bug - create lightning log folders when preparing environment

+++ train.py
@@ -45,7 +45,7 @@ if __name__ == "__main__":
         max_epochs=max_epochs,
         deterministic=True,
         strategy="fsdp",
-        precision=16,
+        precision="16-mixed",
     )
 
     # Auto log all Mlflow entities
